Major refactor of the RayObject hierarchy
RayObject has become more abstractified, now acting as a base for both visible
objects and lights.  The main reason for this is to simplify deserialization;
also, it has become evident that being a light and being a visible object are
not at all exclusive.

On the downside, terrifying multiple-inheritance diamonds are now normal.

Perhaps I should reconsider this...